requirements.txt
flask
flask-mail
yookassa
python-dotenv